+++ changelog.md
@@ -1,5 +1,6 @@
 ## ( [Release](https://github.com/Kevadroz/PracticeCheckpointPermanence/releases/tag/v1.4.5) | [Source](https://github.com/Kevadroz/PracticeCheckpointPermanence/tree/v1.4.5) )
 - Fix Incompatibility with "Labeled Checkpoints" \[[Issue #54](https://github.com/Kevadroz/PracticeCheckpointPermanence/issues/54)\]
+- The "Reset Attempts" option now also resets jumps \[[Issue #49](https://github.com/Kevadroz/PracticeCheckpointPermanence/issues/49)\]
 
 ## ( [Release](https://github.com/Kevadroz/PracticeCheckpointPermanence/releases/tag/v1.4.4) | [Source](https://github.com/Kevadroz/PracticeCheckpointPermanence/tree/v1.4.4) )
 - Disabled iOS Support.